ci: refine release tag automation
